Expose rewriting as library
